+++ ChangeLog.md
@@ -2,6 +2,13 @@
 
 ## [Unreleased]
 
+## [0.1.1.0] - 2019-05-13
+### Added
+- New version of module Bio.Sequence hat introduces `IsSequence` typeclass and `Sequence` datatype.
+- Type `GenBankStructure` describing structure of .gb-file.
+### Changed
+- ABI cleaner and decoder now work with type `ABIRaw`.
+
 ## [0.1.0.1] - 2019-02-28
 ### Added
 - Field `chemCompType` in `Residue`.